fix Approx problem in travis

+++ lib/test/src/test_ck_plane.cpp
@@ -126,8 +126,8 @@ template <typename PG> void chk_float(const PG &myck) {
 
     auto [q1, q2, q3] = myck.tri_quadrance(triangle);
     auto [s1, s2, s3] = myck.tri_spread(trilateral);
-    CHECK( q1 * s2 - q2 * s1 == Approx(0.) );
-    CHECK( q2 * s3 - q3 * s2 == Approx(0.) );
+    CHECK( double(q1 * s2 - q2 * s1) == Approx(0.) );
+    CHECK( double(q2 * s3 - q3 * s2) == Approx(0.) );
     // CHECK( q2 * s3 == Approx(q3 * s2) );
     // CHECK(check_sine_law(Q, S));
     // CHECK(check_sine_law(S, Q));
